% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.r
\name{bin_snps}
\alias{bin_snps}
\title{bin_snps}
\usage{
bin_snps(vcf, bin_size = 100)
}
\arguments{
\item{bin_size}{(int)}

\item{VCF}{in vcfR format}
}
\description{
Extract one random snp in each windows of 'bin' size
}
