% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.r
\name{Convert_GT}
\alias{Convert_GT}
\title{Convert_GT}
\usage{
Convert_GT(GT_table)
}
\arguments{
\item{:}{genotype table, make sure that 'na' are encoded in R readable. In format :

loci_table = data.frame(I=c("0/1" , "0/1" , "1/1", "0/0" , "0/0" , NA, "0/1" , "0/1" , "1/1"),

                       J=c("0/1" , "0/1" , "0/1", NA , "0/0" , NA, "0/1" , "0/1", NA ),

                       K=c("0/1" , "1/1" , "1/1", "0/0" , "0/0" , "1/1", "1/1" , "1/1", "0/0" ))}

\item{vec_pop}{: list of pop in order of genotype table with the following encoding : 

pop_vec=c("A","A","A", "B", "B", "B", "C", "C", "C")

samples=c("s1","s2","s3", "s4", "s5", "s6", "s7", "s8", "s9")

pop_table_test = data.frame(cbind(pop_vec, samples))}
}
\description{
Convert_GT
}
\examples{
loci_table_T = t(loci_table)
colnames(loci_table_T) = samples
# add info on GT table in 0,1,2 genotype codes
loci_table_T_CV = Convert_GT(loci_table_T)
  s1 s2 s3 s4 s5 s6 s7 s8 s9
I  1  1  2  0  0 NA  1  1  2
J  1  1  1 NA  0 NA  1  1 NA
K  1  2  2  0  0  2  2  2  0
}
