% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Filters_VCF.r
\name{filter_vcf_by_dp}
\alias{filter_vcf_by_dp}
\title{filter_vcf_by_dp : Filter Inds and Pos by DP Thresholds}
\usage{
filter_vcf_by_dp(vcf, min_dp_pos, max_dp_pos, min_dp_ind, max_dp_ind)
}
\arguments{
\item{vcf}{A VCF object (vcfR format) to be filtered.}

\item{min_dp_pos}{(double) The minimum allowed DP for positions.}

\item{max_dp_pos}{(double) The maximum allowed DP for positions.}

\item{min_dp_ind}{(double) The minimum allowed DP for individuals.}

\item{max_dp_ind}{(double) The maximum allowed DP for individuals.}
}
\value{
A filtered VCF object where positions and individuals are removed based on the DP thresholds.
}
\description{
This function filters a VCF file based on the depth of coverage (DP) for both positions (pos) and individuals (ind). 
It removes positions and individuals whose DP is outside the specified minimum and maximum thresholds.
}
\examples{
# Filter a VCF file based on DP thresholds for both pos and ind
filtered_vcf <- filter_vcf_by_dp(vcf_file, 10, 100, 20, 200)
}
