% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Filters_VCF.r
\name{Filters_na_Ind_Pos}
\alias{Filters_na_Ind_Pos}
\title{Filters_na_Ind_Pos : Filters VCF based on missing data rate for individuals and positions}
\usage{
Filters_na_Ind_Pos(vcf, rate_na_max_POS, rate_na_max_Ind, ploidy)
}
\arguments{
\item{vcf}{A VCF object (vcfR object) to be filtered.}

\item{rate_na_max_POS}{Maximum allowed percentage of missing data per position.}

\item{rate_na_max_Ind}{Maximum allowed percentage of missing data per individual.}

\item{ploidy}{Ploidy level (1 or 2) to determine the missing data representation (`.` or `./.`).}
}
\value{
A filtered VCF object.
}
\description{
This function filters a VCF object by the missing data rate for both individuals (samples)
and positions (variants). The user can specify the ploidy level (1 or 2), and the function
will compute the missing data rate based on either the `.` (for ploidy 1) or `./.` (for ploidy 2)
genotype representation.
}
