% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_SFS.R
\name{Get2DSFS}
\alias{Get2DSFS}
\title{Get2DSFS: Compute pairwise Hudson Fst on whole loci}
\usage{
Get2DSFS(pop_table, pop_pair, loci_table_T)
}
\arguments{
\item{pop_table}{Data.frame with pop_vec and sample names}

\item{loci_table_T}{Transposed genotype table with 0,1,2 encoding}

\item{pop_pairs}{Pair of pop names}
}
\value{
count matrix of 10 class allele frequencies

 @description
 Discard all position with NA or monomorphic in both pops
 Get the minor allele frequencies by pop as following : 
     # count of minor and major allele
     count_snp_i=as.numeric(rowSums(current_pop_i))
     count_ref_i=as.numeric(size_pop_i*2 - rowSums(current_pop_i))
     # get minor allele count
     minor_count_i = pmin(count_ref_i,count_snp_i)
     # get minor allele freq
     minor_freq_i = minor_count_i / size_pop_i
     # assign class to freq
     class_i_vec <- findInterval(minor_freq_i, breaks)
     freq_pop[[pop]] = class_i_vec
}
\description{
Get2DSFS: Compute pairwise Hudson Fst on whole loci
}
\examples{
:

loci_table_test=loci_table_convert[c(1:50000),, drop = FALSE]

metadata table : 
     Population   GT_sample
2     pop1 s1
3     pop1 s2
4     pop2 s5
5     pop2 s6 ....
      .......
     
SFS_2D = Get2DSFS(pop_table = metadata, pop_pair = c("pop1", "pop2"), loci_table_T = loci_table_convert)

[1] "pop sizes"
[1]  12 11

> SFS_2D
         1      2      3      4      5      6      7      8      9     10     11
 1  118068   1741   1327    986    977    872    871    850    795    788    388
 2   10934    197    156    141    147    165    139    122    144    111     51
 3    7515    151    169    118    119    130    142    115    122    130     69
 4    5203    169    153    115    127    121    130    116    124    102     44
 5    3931    160    134    123    145    150    132    148    107    134     67
 6    5432    319    292    245    267    250    238    256    229    214    120
 7    2026    155    140    108    128    119    125    116    118    119     49
 8    1808    114    139    125    123     91    123    102     97    112     56
 9    1623    157    153    110    136    106    102     93     93    130     51
 10   1525    129    135     96    130    118    115    103     96     94     73
 11    771     66     68     72     67     55     59     46     45     46     40
}
