% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_SFS.R
\name{Get_FoldedSFS}
\alias{Get_FoldedSFS}
\title{Get_FoldedSFS}
\usage{
Get_FoldedSFS(loci_table_T)
}
\arguments{
\item{loci_table_T}{Transposed genotype table with 0,1,2 encoding}
}
\value{
vector of counts of minor allele frequencies
}
\description{
Get_FoldedSFS
}
\examples{
:
Compute folded SFS as following : 
  count_snp=as.numeric(sum(current_pos))
  count_ref=as.numeric(nb_samples*2 - count_snp)
  minor_count = pmin(count_ref,count_snp)

loci_table_Test=loci_table_convert[c(1:50000),]
dim(loci_table_Test)
SFS_folded=Get_FoldedSFS(loci_table_Test)
[1] "nb_class:"
[1] 31
SFS_folded
[1]     1     2     3     4     5     6     7     8     9    10    11    12    13    14    15    16    17    18    19    20    21    22    23    24    25    26    27    28    29 
#     25810 12912  8908  6501  5006  3969  3148  2772  2340  2079  1934  1724  1473  1450  1400  1325  1221  1339  1203  1115  1137  1225  1090  1054  1039  1038  1082   971   962 
}
