% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.r
\name{Get_SW_abs}
\alias{Get_SW_abs}
\title{Get_SW_abs}
\usage{
Get_SW_abs(VCF, slide = 5000, window = 10000)
}
\arguments{
\item{window}{window (int)}

\item{:}{VCF : vcfR format}

\item{slides}{slide (int)}
}
\value{
a dataframe of sliding windows with colnames = c(chr,start,end, mid_point, nb_snp)
}
\description{
uses Granges and Iranges object to create sw windows table
}
\examples{
df = Get_SW_abs(VCFR_data, slide = 2000, windows = 10000)
df
 chr        start   end mid_point nb_snp
<chr>      <int> <int>     <dbl>  <int>
1 ptg000007l     1 10000     5000.     21
2 ptg000007l  2001 12000     7000.     60
3 ptg000007l  4001 14000     9000.     73
4 ptg000007l  6001 16000    11000.     87
5 ptg000007l  8001 18000    13000.    109
6 ptg000007l 10001 20000    15000.     92
7 ptg000007l 12001 22000    17000.     61
8 ptg000007l 14001 24000    19000.     55
9 ptg000007l 16001 26000    21000.     49
10 ptg000007l 18001 28000    23000.     44
6,915 more rows
}
