% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.r
\name{Get_SW_SNPrange}
\alias{Get_SW_SNPrange}
\title{Get_SW_SNPrange}
\usage{
Get_SW_SNPrange(VCF, nb_snp_wind = 1000, nb_snp_slide = 500)
}
\arguments{
\item{nb_snp_wind}{window (int)}

\item{nb_snp_slide}{slide (int)}

\item{:}{VCF : vcfR format}
}
\value{
a dataframe of sliding windows with colnames = c(chr,start,end, mid_point)
}
\description{
uses Granges and Iranges object to create sw windows table over SNPs
}
\examples{
df = Get_SW_SNPrange(VCFR_data, nb_snp_wind = 2000, nb_snp_slide = 1000)
df
403 × 4
chr        start   end mid_point
<chr>      <int> <int>     <dbl>
 1 ptg000007l     1  1000      500.
2 ptg000007l   501  1500     1000.
3 ptg000007l  1001  2000     1500.
4 ptg000007l  1501  2500     2000.
5 ptg000007l  2001  3000     2500.
6 ptg000007l  2501  3500     3000.
7 ptg000007l  3001  4000     3500.
8 ptg000007l  3501  4500     4000.
9 ptg000007l  4001  5000     4500.
10 ptg000007l  4501  5500     5000.
}
