% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.r
\name{getAlleleCount}
\alias{getAlleleCount}
\title{getAltCount : get the Alt and Ref allele count in genotype table (0,1,2 encoded)}
\usage{
getAlleleCount(locus)
}
\arguments{
\item{:}{locus table, make sure that 'NA' are encoded in R readable.}
}
\value{
List of alt allele count in the pop.
}
\description{
The na are removed from the count calculation, so the number of total allele differs between SNPs
}
\examples{
getAlleleFreq(locus)
$alt_freq
[1] 0.4

$ref_freq
[1] 0.6
}
