% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.r
\name{getAlleleCountByPop}
\alias{getAlleleCountByPop}
\title{getAlleleCountByPop : Extract allele count from genotype table by pop use getAlleleCount function}
\usage{
getAlleleCountByPop(loci_table, pop_table)
}
\arguments{
\item{loci_table}{dataframe of allele count in loci (rows) and in samples (col) from one or several pop}

\item{pop_table}{(dataframe) : contains samples ID in col1 and Pop names in col2}
}
\value{
list of allele count for each pop
}
\description{
getAlleleCountByPop : Extract allele count from genotype table by pop use getAlleleCount function
}
\examples{
> locus_table
A tibble: 181,043 × 31
  `1622W1_S254` `1623W1_S276` `1624W1_S392` `1625Q_S431` `1625W1_S384` `1626W1_S364` `1627W1_S385` `1629W1_S379` `1681W1_S374` `1682W1_S380` `1684W1_S376` `1686W1_S402`
1             0             0             0            0             0             0             0             0             0             0             0             0
2             0             0             0            0             0             0             0             0             0             0             0             0
3             0             0             0            0             0             0             0             0             0             0             0             0
> pop_table_test
pop_vec samples
1       A      s1
2       A      s2
3       A      s3
4       B      s4
5       B      s5
6       B      s6
7       C      s7
8       C      s8
9       C      s9
by_popCount = getAlleleCountByPop(loci_table = loci_table_T_CV, pop_table = metadata)
by_popCount[[1]]$alt_count
by_popCount$Pop1$ref_count
}
