% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.r
\name{getAlleleFreqByPop}
\alias{getAlleleFreqByPop}
\title{getAlleleFreqByPop : Extract pair of pop in a metadata table}
\usage{
getAlleleFreqByPop(loci_table, pop_table)
}
\arguments{
\item{loci_table}{dataframe of allele count in loci (rows) and in samples (col) from one or several pop}

\item{pop_table}{(dataframe) : contains samples ID in col1 and Pop names in col2}
}
\value{
list of allele frequencies for each pop
}
\description{
getAlleleFreqByPop : Extract pair of pop in a metadata table
}
\examples{
> locus_table
s1 s2 s3 s4 s5 s6 s7 s8 s9
I  1  1  2  0  0 NA  1  1  2
> pop_table_test
pop_vec samples
1       A      s1
2       A      s2
3       A      s3
4       B      s4
5       B      s5
6       B      s6
7       C      s7
8       C      s8
9       C      s9
> summary(res)
        Length Class  Mode
pop1 2      -none- list
pop2   2      -none- list
pop3 2      -none- list
> summary(res$pop1)
        Length Class  Mode   
alt_freq 181043 -none- numeric
ref_freq 181043 -none- numeric
}
