% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_HudsonFst.r
\name{getFstBySNP_Count}
\alias{getFstBySNP_Count}
\title{getFstBySNP_Count : Compute Pi based on alt count}
\usage{
getFstBySNP_Count(loci_table_T, pop_table, Na_rate = 0.2, MAF_threshold = 0.05)
}
\arguments{
\item{loci_table_T}{: dataframe format with 0,1,2 encoded genotype, samples in column and position in lines}

\item{pop_table}{: data.frame with col1 : pop (char) and col2 : ind (car)}

\item{Na_rate}{: na rate}

\item{MAF_threshold}{: MAF}
}
\value{
vec_pi_pop : vector of Pi within and Pi between in  2 lists format
}
\description{
getFstBySNP_Count : Compute Pi based on alt count
}
\examples{
: 

# get genotype table
loci_table = extract.gt(VCFR_data, element = "GT")
loci_table = as.data.frame(loci_table)
geno_table = Convert_GT(GT_table = loci_table)

geno_table[c(1:10),c(1:10)]
ptg000007l_8628           0           0           0          0           0           0           0           0           0
ptg000007l_8631           0           0           0          0           0           0           0           0           0
ptg000007l_8633           0           0           0          0           0           0           0           0           0
ptg000007l_8639           0           0           0          0           0           0           0           0           2
ptg000007l_9618          NA          NA          NA         NA           1           1          NA          NA           1
ptg000007l_9622          NA          NA           0         NA           1           1          NA          NA           1
ptg000007l_9625          NA          NA           0         NA           1           1          NA          NA           1
ptg000007l_9626          NA          NA           0         NA           1           1          NA          NA           1
Fst_by_SNP = getFstBySNP_Count(loci_table_T = geno_table, pop_table = metadata_sorted, Na_rate = 0.20, MAF_threshold = 0.05)
summary(Fst_by_SNP)
> summary(Fst_by_SNP)
Length Class  Mode   
Termes_Cesseras             17785 -none- numeric
Termes_Cesseras_position   201634 -none- numeric
Termes_Cucugnan             54847 -none- numeric
Termes_Cucugnan_position   201634 -none- numeric
Cesseras_Cucugnan           24931 -none- numeric
Cesseras_Cucugnan_position 201634 -none- numeric
}
