% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_Genotype_Freq.R
\name{GetGenoFreqByPop}
\alias{GetGenoFreqByPop}
\title{GetGenoFreqByPop}
\usage{
GetGenoFreqByPop(geno_table, chr_name = NULL, ploidy = NULL, pop_list)
}
\arguments{
\item{geno_table}{: dataframe format with 0/1 diploid genotype format or 0 ,1 haplpoid genotype, samples in column and position in lines}

\item{pop_list}{: list of samples (values) in each pop (names)}
}
\value{
geno_freq : table with  col : "chr", "position", "POP1_het_freq", "POP1__hom_alt_freq", "POP1_hom_ref_freq"
}
\description{
Na are removed from the genotype count but not the corresponding sample. The function should be run on table with no missing for more accuracy
}
\examples{
: 
loci_table = extract.gt(VCFR_data, element = "GT")
loci_table = as.data.frame(loci_table)
loci_table[c(1:10),c(1:10)]
            1622W1_S254 1623W1_S276 1624W1_S392 1625Q_S431 1625W1_S384 1626W1_S364 1627W1_S385 1629W1_S379 1681W1_S374 1682W1_S380
ptg000007l_9661          0/0         0/0         0/0        0/0         0/0         0/0         0/0         0/0         0/0         0/0
ptg000007l_9668          0/0         0/0         0/0        0/0         0/0         0/0         0/0         0/0         0/0         0/0
ptg000007l_9670          0/0         0/0         0/0        0/0         0/0         0/0         0/0         0/0         0/0         0/0
ptg000007l_9678          0/0         0/0         0/0        0/0         0/0         0/0         0/0         0/0         1/1         1/1

pop_list = pop_list
$POP1
[1] "1681W1_S374" "1682W1_S380" "1684W1_S376" "1686W1_S402" 
$POP2
[1] "1551W1_S395" "1553W1_S314" "1554W1_S341" "1555W1_S397" "1556W1_S337" 
$POP3
[1] "1622W1_S254" "1623W1_S276" "1624W1_S392" "1625Q_S431"  "1625W1_S384"

table = GetGenoFreqByPop(geno_table = loci_table, chr_name = chr, ploidy = "diploid", pop_list = pop_list)

chr position POP1_het_freq POP1_hom_alt_freq POP1_hom_ref_freq POP1_hom_tot_freq POP1_het_freq ... same with POP2 and POP3
1  ptg000007l     9661         0.0000000             0.0000000             1.0000000             1.0000000        0.00000000
2  ptg000007l     9668         0.0000000             0.0000000             1.0000000             1.0000000        0.00000000
3  ptg000007l     9670         0.0000000             0.0000000             1.0000000             1.0000000        0.08333333
4  ptg000007l     9678         0.2727273             0.4545455             0.2727273             0.7272727        0.00000000
5  ptg000007l     9687         0.0000000             0.0000000             1.0000000             1.0000000        0.00000000
}
