% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_Fst.R
\name{getGlobalFst}
\alias{getGlobalFst}
\title{Compute pairwise Hudson Fst on whole loci}
\usage{
getGlobalFst(loci_table_T, pop_table, Na_rate = 0.2, MAF_threshold = 0.05)
}
\arguments{
\item{loci_table_T}{Transposed genotype table with 0,1,2 encoding}

\item{pop_table}{Data.frame with pop_vec and sample names}

\item{Na_rate}{= 0.2}

\item{MAF_threshold}{= 0.05}
}
\value{
List of Fst values and population pair
}
\description{
Compute pairwise Hudson Fst on whole loci
}
\examples{
 : run Global Hudson Fst 

# input : genotype table, make sure that na are encoded in upper case, R readable. In forme :
loci_table = data.frame(I=c("0/1" , "0/1" , "1/1", "0/0" , "0/0" , NA, "0/1" , "0/1" , "1/1"),
                      J=c("0/1" , "0/1" , "0/1", NA , "0/0" , NA, "0/1" , "0/1", NA ),
                      K=c("0/1" , "1/1" , "1/1", "0/0" , "0/0" , "1/1", "1/1" , "1/1", "0/0" ))

# vec_pop : list of pop in order of genotype table with the following encoding : 
pop_vec=c("A","A","A", "B", "B", "B", "C", "C", "C")
samples=c("s1","s2","s3", "s4", "s5", "s6", "s7", "s8", "s9")
pop_table_test = data.frame(cbind(pop_vec, samples))

loci_table_T = t(loci_table)
colnames(loci_table_T) = samples

# add info on GT table in 0,1,2 genotype codes
loci_table_T_CV = Convert_GT(loci_table_T)

# get hudson Fst by pop pairs written in pop_table_test
getGlobalFst(loci_table_T = loci_table_T_CV, pop_table = pop_table_test)
$A_B
[1] 0.34375

$A_C
[1] -0.4375

$B_C
[1] 0.3709677
}
