% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_HudsonFst.r
\name{getGlobalFst_Count}
\alias{getGlobalFst_Count}
\title{getGlobalFst_Count : Compute pairwise Hudson Fst on ALT count}
\usage{
getGlobalFst_Count(
  loci_table_T,
  pop_table,
  Na_rate = 0.2,
  MAF_threshold = 0.05
)
}
\arguments{
\item{loci_table_T}{Transposed genotype table with 0,1,2 encoding}

\item{pop_table}{Data.frame with pop_vec and sample names}

\item{Na_rate}{= 0.2}

\item{MAF_threshold}{= 0.05}
}
\value{
List of Fst values and population pair
}
\description{
getGlobalFst_Count : Compute pairwise Hudson Fst on ALT count
}
\examples{

`1622W1_S254` `1623W1_S276` `1624W1_S392` `1625Q_S431` `1625W1_S384` `1626W1_S364` `1627W1_S385` `1629W1_S379` `1681W1_S374` `1682W1_S380` `1684W1_S376` `1686W1_S402`
1             0             0             0            0             0             0             0             0             0             0             0             0
2             0             0             0            0             0             0             0             0             0             0             0             0
3             0             0             0            0             0             0             0             0             0             0             0             0
# ℹ 19 more variables: `
> metadata
  Population   GT_sample
1    pop1 1551W1_S395
2    pop1 1553W1_S314
3    pop1 1554W1_S341
4    pop1 1555W1_S397
5    pop1 1556W1_S337
Fst_GlobalCount = getGlobalFst_Count(loci_table_T = loci_table_T_CV, 
                         pop_table = meta_sub, Na_rate = 0.2, MAF_threshold = 0)
					  tibble(loci_table_T_CV)
[1] "pop1" "pop2"  
[1] "get count:"
[1] "pop1"
[1] "1551W1_S395" "1553W1_S314" "1554W1_S341" "1555W1_S397" "1556W1_S337" "1557Q_S432"  "1557W1_S275" "1558W1_S354" "1559W1_S357" "1560W1_S300" "1562W1_S242" "1563W1_S281"
[1] "get count:"
[1] "pop2"
[1] "1622W1_S254" "1623W1_S276" "1624W1_S392" "1625Q_S431"  "1625W1_S384" "1626W1_S364" "1627W1_S385" "1629W1_S379"
[1] "nb of loci : "
[1] 54200
[1] 0.2316629
}
