% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_Fst.R
\name{getPiWithinBetween}
\alias{getPiWithinBetween}
\title{Compute Pi on loci}
\usage{
getPiWithinBetween(loci_pairs, pop_table)
}
\arguments{
\item{pop_table}{: data.frame with col1 : pop (char) and col2 : ind (car)}

\item{loci}{: dataframe format with 0,1,2 encoded genotype, samples in column and ind in lines}
}
\value{
vec_pi_pop : vector of Pi within and Pi between for the loci
}
\description{
Compute Pi on loci
}
\examples{
: run Pi computation

# input : genotype table, make sure that na are encoded in upper case, R readable. In forme :
loci_table = data.frame(I=c("0/1" , "0/1" , "1/1", "0/0" , "0/0" , NA, "0/1" , "0/1" , "1/1"),
                      J=c("0/1" , "0/1" , "0/1", NA , "0/0" , NA, "0/1" , "0/1", NA ),
                      K=c("0/1" , "1/1" , "1/1", "0/0" , "0/0" , "1/1", "1/1" , "1/1", "0/0" ))

# vec_pop : list of pop in order of genotype table with the following encoding : 
pop_vec=c("A","A","A", "B", "B", "B", "C", "C", "C")
samples=c("s1","s2","s3", "s4", "s5", "s6", "s7", "s8", "s9")
pop_table_test = data.frame(cbind(pop_vec, samples))

loci_table_T = t(loci_table)
colnames(loci_table_T) = samples

# add info on GT table in 0,1,2 genotype codes
loci_table_T_CV = Convert_GT(loci_table_T)

# compute Pi :
getPiWithinBetween(loci = loci_test, pop_table = pop_table_test) 
[1] 0.3333333 0.6666667 
attr(,"levels") 
[1] "Pi_within" "Pi_between"
}
