% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_HudsonFst.r
\name{getPiWithinBetween_Count}
\alias{getPiWithinBetween_Count}
\title{getPiWithinBetween_Count : Compute Pi based on alt count}
\usage{
getPiWithinBetween_Count(loci_pairs, pop_table)
}
\arguments{
\item{loci_pairs}{: dataframe format with 0,1,2 encoded genotype, samples in column and positions in lines}

\item{pop_table}{: data.frame with col1 : pop (char) and col2 : ind (car)}
}
\value{
vec_pi_pop : vector of Pi within and Pi between in  2 lists format
}
\description{
getPiWithinBetween_Count : Compute Pi based on alt count
}
\examples{
: TODO
}
