% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.r
\name{getSamplePairPop}
\alias{getSamplePairPop}
\title{getSamplePairPop : Extract pair of pop in a metadata table}
\usage{
getSamplePairPop(Pair_i, pop_table)
}
\arguments{
\item{Pair_i}{vector of 2 pop names (char)}

\item{pop_table}{(dataframe) : contains samples ID in col1 and Pop names in col2}
}
\value{
list of samples named by pop
}
\description{
getSamplePairPop : Extract pair of pop in a metadata table
}
\examples{
 > pop_table_test
pop_vec samples
1       A      s1
2       A      s2
3       A      s3
4       B      s4
5       B      s5
6       B      s6
7       C      s7
8       C      s8
9       C      s9
> getSamplePairPop(Pair_i = c("A", "B"), pop_table_test)
$A
[1] "s1" "s2" "s3"

$B
[1] "s4" "s5" "s6"
}
