% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{getSamplePop}
\alias{getSamplePop}
\title{getSamplePop : Extract one of pop in a metadata table}
\usage{
getSamplePop(pop_name, pop_table)
}
\arguments{
\item{pop_name}{(char)}

\item{pop_table}{(dataframe) : contains samples ID in col1 and Pop names in col2}
}
\value{
list of samples named by pop
}
\description{
getSamplePop : Extract one of pop in a metadata table
}
\examples{
> pop_table_test
pop_vec samples
1       A      s1
2       A      s2
3       A      s3
4       B      s4
5       B      s5
6       B      s6
7       C      s7
8       C      s8
9       C      s9
> getSamplePop(pop_name = "A", pop_table_test)
[1] "s1" "s2" "s3"
attr(,"levels")
[1] "A"
}
