% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{PurgeRefInPops}
\alias{PurgeRefInPops}
\title{PurgeRefInPops : detect and discard homozygous loci (usefull for pop subset table)}
\usage{
PurgeRefInPops(loci_table, pop_table)
}
\arguments{
\item{loci_table}{: genotype table, make sure that 'na' are encoded in R readable. In format :

loci_table (df) : 

    s1 s2 s3 s4 s5 s6 s7 s8 s9

I  1  1  2  0  0 NA  1  1  2

J  1  1  1 NA  0 NA  1  1 NA

K  1  2  2  0  0  2  2  2  0

L  0  0  0  0  0  0  0  0  0}

\item{pop_table}{: list of pop in order of genotype table with the following encoding : 

pop_vec=c("A","A","A", "B", "B", "B", "C", "C", "C")

samples=c("s1","s2","s3", "s4", "s5", "s6", "s7", "s8", "s9")

pop_table_test = data.frame(cbind(pop_vec, samples))}
}
\value{
subset of loci table
}
\description{
PurgeRefInPops : detect and discard homozygous loci (usefull for pop subset table)
}
\examples{
loci_table_T = t(loci_table)
colnames(loci_table_T) = samples
loci_table_T_CV = Convert_GT(loci_table_T)
SubSnp = PurgeRefInPops(loci_table = loci_table_T_CV_ex, pop_table = pop_table_test)
  s1 s2 s3 s4 s5 s6 s7 s8 s9
I  1  1  2  0  0 NA  1  1  2
J  1  1  1 NA  0 NA  1  1 NA
K  1  2  2  0  0  2  2  2  0

}
