% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Filters_VCF.r
\name{Filters_Het_Ind_Pos}
\alias{Filters_Het_Ind_Pos}
\title{Filter Individuals and Positions by Alternative Rate}
\usage{
Filters_Het_Ind_Pos(vcf, ploidy, rate_het_max_POS, rate_het_max_Ind)
}
\arguments{
\item{ploidy}{(integer) The ploidy level of the dataset. Either 1 or 2.}

\item{rate_het_max_POS}{(double) The maximum allowed alternative allele rate (percent) for positions.}

\item{rate_het_max_Ind}{(double) The maximum allowed alternative allele rate (percent) for individuals.}

\item{vfc}{A VCF object (vcfR format) to be filtered.}

\item{rate_het_min_pos}{(integer) The minimum number of alternative alleles allowed per position.}
}
\value{
A filtered VCF object where positions and individuals are removed based on the alternative allele rate filters.
}
\description{
This function filters a VCF file based on the alternative allele rate both by individuals and by positions. 
The function calculates the percentage of heterozygous (alternative) genotypes for each individual and for each position,
and filters the VCF file according to the provided thresholds. The function also handles ploidy (1 or 2) by adjusting
the genotypes accordingly ("0/1" for ploidy 2, "1" for ploidy 1).
}
\examples{
# Filter a VCF file based on alternative rate thresholds for ploidy 2
filtered_vcf <- Filters_Het_Ind_Pos(vcf_file, 2, 30.0, 50.0, 0.0)

# Filter a VCF file based on alternative rate thresholds for ploidy 1
filtered_vcf <- Filters_Het_Ind_Pos(vcf_file, 1,  30.0, 50.0, 0.0)
}
