% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_SFS.R
\name{Get2DSFS}
\alias{Get2DSFS}
\title{Get2DSFS: Compute pairwise Hudson Fst on whole loci}
\usage{
Get2DSFS(pop_table, pop_pair, loci_table_T)
}
\arguments{
\item{pop_table}{Data.frame with pop_vec and sample names}

\item{loci_table_T}{Transposed genotype table with 0,1,2 encoding}

\item{pop_pairs}{Pair of pop names}
}
\value{
count matrix of 10 class allele frequencies

 @description
 Discard all position with NA or monomorphic in both pops
 Get the minor allele frequencies by pop as following : 
   count_snp_i=as.numeric(sum(current_pop_i))
   count_ref_i=as.numeric(size_pop_i*2 - count_snp_i)
   # get minor allele count
   minor_count_i = min(count_ref_i,count_snp_i)
   # get minor allele freq
   minor_freq_i = minor_count_i / size_pop_i
}
\description{
Get2DSFS: Compute pairwise Hudson Fst on whole loci
}
\examples{
:

loci_table_test=loci_table_convert[c(1:50000),, drop = FALSE]

metadata table : 
     Population   GT_sample
2     pop1 s1
3     pop1 s2
4     pop2 s5
5     pop2 s6 ....
      .......
     
SFS_2D = Get2DSFS(pop_table = metadata, pop_pair = c("pop1", "pop2"), loci_table_T = loci_table_convert)

[1] "pop sizes"
[1]  8 11

> SFS_2D
c(0, 0.1) c(0.1, 0.2) c(0.2, 0.3) c(0.3, 0.4) c(0.4, 0.5) c(0.5, 0.6) c(0.6, 0.7) c(0.7, 0.8) c(0.8, 0.9) c(0.9, 1)
c(0, 0.1)        4742        1567        1150         826         825         627         658         585         535       585
c(0.1, 0.2)      7010         350         307         245         259         257         277         263         257       218
c(0.2, 0.3)      4168         260         213         202         240         250         213         239         216       202
c(0.3, 0.4)      3110         223         219         170         198         186         190         194         202       204
c(0.4, 0.5)         0           0           0           0           0           0           0           0           0         0
c(0.5, 0.6)      2739         228         226         203         195         198         191         192         186       182
c(0.6, 0.7)      2367         205         201         157         160         186         184         191         164       166
c(0.7, 0.8)      2371         215         218         162         208         199         210         153         148       187
c(0.8, 0.9)      1923         225         224         167         185         176         166         165         171       163
c(0.9, 1)           0           0           0           0           0           0           0           0           0         0
}
