% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_SFS.R
\name{Get_FoldedSFS}
\alias{Get_FoldedSFS}
\title{Get2DSFS: Compute pairwise Hudson Fst on whole loci}
\usage{
Get_FoldedSFS(loci_table_T)
}
\arguments{
\item{loci_table_T}{Transposed genotype table with 0,1,2 encoding}
}
\value{
vector of counts of minor allele frequencies
}
\description{
Get2DSFS: Compute pairwise Hudson Fst on whole loci
}
\examples{
:
Compute folded SFS as following : 
  count_snp=as.numeric(sum(current_pos))
  count_ref=as.numeric(nb_samples*2 - count_snp)
  minor_count = min(count_ref,count_snp)

loci_table_Test=loci_table_convert[c(1:50000),]
dim(loci_table_Test)
SFS_folded=Get_FoldedSFS(loci_table_Test)
[1] "nb_class:"
[1] 31
unlist(SFS_folded)
[1] 6181 3751 2238 1773 1373 1034  789  765  674  537  444  399  393  350  298  298  285  265  246  211  212  214  179  192  181  171  167  138  149  152  109
}
