% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{getAlleleFreq}
\alias{getAlleleFreq}
\title{getAlleleFreq : get the Alt and Ref allele frequencies in one locus table}
\usage{
getAlleleFreq(locus)
}
\arguments{
\item{:}{locus table of one line, make sure that 'NA' are encoded in R readable. 

In format : Obtain by convert_GT()

  s1 s2 s3 s4 s5 s6 s7 s8 s9

I  1  1  2  0  0 NA  1  1  2}
}
\value{
List of alt and ref allele frequencies in the pop. Na are removed from the sample number
}
\description{
getAlleleFreq : get the Alt and Ref allele frequencies in one locus table
}
\examples{
getAlleleFreq(locus)
$alt_freq
[1] 0.4

$ref_freq
[1] 0.6
}
