% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{getPairwisePop}
\alias{getPairwisePop}
\title{getSamplePop : Extract one of pop in a metadata table}
\usage{
getPairwisePop(pop_vec)
}
\arguments{
\item{pop_vec}{(char  vec)}
}
\value{
df of pairwise pop couple (by row)
}
\description{
getSamplePop : Extract one of pop in a metadata table
}
\examples{
pairs_pop_sample = getPairwisePop(unique(pop_table_test[,1]))
}
